[refactor] DetailVC와 DetailView를 분리
