+++ Soulmate/Soulmate/Presentation/HomeScene/View/DetailView.swift
@@ -1,0 +1,136 @@
+//
+//  DetailView.swift
+//  Soulmate
+//
+//  Created by Sangmin Lee on 2022/12/19.
+//
+
+import UIKit
+import Combine
+import SnapKit
+
+final class DetailView: UIView {
+    @Published var currentImagePage: Int = 0
+
+    lazy var collectionView: UICollectionView = {
+        let layout = createCompositionalLayout()
+        let collection = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        collection.backgroundColor = .systemBackground
+        collection.bounces = true
+        collection.showsVerticalScrollIndicator = false
+        collection.showsHorizontalScrollIndicator = false
+        self.addSubview(collection)
+        return collection
+    }()
+    
+    lazy var applyButton: HeartConsumeGradientButton = {
+        let button = HeartConsumeGradientButton(title: "대화친구 신청하기")
+        self.addSubview(button)
+        return button
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        configureView()
+        configureLayout()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func configureView() {
+        self.backgroundColor = .systemBackground
+    }
+    
+    private func configureLayout() {
+        collectionView.snp.makeConstraints {
+            $0.leading.trailing.top.equalTo(self.safeAreaLayoutGuide)
+            $0.bottom.equalTo(applyButton.snp.top)
+        }
+        
+        applyButton.snp.makeConstraints {
+            $0.leading.trailing.equalTo(self.safeAreaLayoutGuide).inset(20)
+            $0.bottom.equalTo(self.snp.bottom).inset(46)
+            $0.height.equalTo(54)
+        }
+    }
+}
+
+extension DetailView {
+    private func createCompositionalLayout() -> UICollectionViewCompositionalLayout {
+        return UICollectionViewCompositionalLayout { sectionNumber, _ -> NSCollectionLayoutSection? in
+            switch sectionNumber {
+            case 0: return self.photoLayoutSection()
+            case 1: return self.profileLayoutSection()
+            case 2: return self.greetingLayoutSection()
+            case 3: return self.basicInfoLayoutSection()
+            default: return self.basicInfoLayoutSection()
+            }
+        }
+    }
+    
+    private func photoLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(1))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        group.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
+        
+        let section = NSCollectionLayoutSection(group: group)
+        section.orthogonalScrollingBehavior = .groupPaging
+        
+        let footerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(40))
+        let footer = NSCollectionLayoutBoundarySupplementaryItem(
+            layoutSize: footerSize,
+            elementKind: PhotoFooterView.footerKind,
+            alignment: .bottom
+        )
+    
+        section.boundarySupplementaryItems = [footer]
+        section.visibleItemsInvalidationHandler = { [weak self] _, offset, _ -> Void in
+            guard let width = self?.bounds.width else { return }
+            let page = round(offset.x / width)
+            self?.currentImagePage = Int(page)
+        }
+        return section
+    }
+    
+    private func profileLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(0.3))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(0.3))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        return section
+    }
+    
+    private func greetingLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(100))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        return section
+    }
+    
+    private func basicInfoLayoutSection() -> NSCollectionLayoutSection {
+        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(0.5))
+        let item = NSCollectionLayoutItem(layoutSize: itemSize)
+        item.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
+        
+        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(0.5))
+        let group = NSCollectionLayoutGroup.horizontal(layoutSize: groupSize, subitems: [item])
+        
+        let section = NSCollectionLayoutSection(group: group)
+        return section
+    }
+}
